Infinity_FinTech_Backend/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-admin==2.0.2
django-cors-headers==4.7.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.16.1
excel-base==1.0.4
gunicorn==23.0.0
isoweek==1.3.3
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
screen==1.0.1
six==1.17.0
sqlparse==0.5.3
TimeConvert==3.0.13
tzdata==2025.2
tzlocal==5.3.1
whitenoise==6.9.0
xlwt==1.3.0
